% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarCoordinatesGadget.R
\name{StarCoordinates}
\alias{StarCoordinates}
\title{Star Coordinates Gadget}
\usage{
StarCoordinates(
  df,
  colorVar = NULL,
  approach = "Standard",
  numericRepresentation = TRUE,
  meanCentered = TRUE,
  projMatrix = NULL,
  clusterFunc = NULL
)
}
\arguments{
\item{df}{A dataframe with the data to explore. It should contain only numeric or factor columns.}

\item{colorVar}{column where labels from the data are extracted.}

\item{approach}{Standard approach as defined by Kandogan, or Orthographic Star Coordinates (OSC) with a recondition as defined by Lehmann and Thiesel}

\item{numericRepresentation}{if true attempt to convert all factors to numeric representation, otherwise used mixed representation as defined in Hinted Star Coordinates}

\item{meanCentered}{center the projection at the mean of the values. May allow for easier value estimation}

\item{projMatrix}{a pre-defined projection matrix as an initial configuration. Should be defined in the same fashion as the output}

\item{clusterFunc}{function to define hints, assume increase in value of the function is an increase in quality of the projection. The function will be called with two parameters (points, labels)}
}
\value{
A list with the projection matrix, coordinates of the projected samples and a logical vector with the selected samples
}
\description{
Creates a RShiny Gadget for Star Coordinates
}
\examples{
library(datasets)
data(iris)
# StarCoordinates(iris, "Species")
}
